% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_parameters}
\alias{extract_parameters}
\title{Extract parameter estimates from SeroCOP fit}
\usage{
extract_parameters(serocop_obj, prob = 0.95)
}
\arguments{
\item{serocop_obj}{A fitted SeroCOP object}

\item{prob}{Probability for credible intervals (default 0.95)}
}
\value{
Data frame with parameter estimates and credible intervals
}
\description{
Extract posterior summaries of model parameters
}
