% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroCOP-package.R
\docType{package}
\name{seroCOP-package}
\alias{seroCOP}
\alias{seroCOP-package}
\title{seroCOP: Correlates of Protection Analysis Using Bayesian Methods}
\description{
Fits generalized logistic functions to serological data (antibody titres) and infection outcomes using Bayesian methods implemented in Stan. Provides tools for assessing correlates of risk/protection including ROC AUC, LOO-CV, and Brier scores. Built around an R6 class for streamlined workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://seroanalytics.org/seroCOP/}
  \item Report bugs at \url{https://github.com/seroanalytics/seroCOP/issues}
}

}
\author{
\strong{Maintainer}: David Hodgson \email{david.hodgson@example.com}

}
\keyword{internal}
